==> .ABGABE ORDNER Steffen/Programmcode/4_lineares_model_input.Rmd <==
---
title: "linearer_model_input"
output: html_notebook
---
Tobias
überarbeitet am 24.01.2020

#title: "linearer model input
#Ziel: model mit den master df variablen füttern
#Autor: "Tobias Lindenau"
#Datum des Erstellens: 18.01.2020
#Aenderungen von Tobias


```{r}
#install.packages("Metrics")
#install.packages("hrbrthemes")
#Bibliotheken
library(readr)
library(tidyverse)
library(lubridate)
library(ggplot2)
library(dplyr)
library(magrittr)
library(broom)
library(Metrics)
library(hrbrthemes)

```

Einlesen der Daten
```{r}
# master df einlesen
master_df <- read.csv("master_df.csv")

master_df$Datum <- as.Date(master_df$Datum, format = "%Y-%m-%d")

#X Spalte loeschen
master_df$X<-NULL

```

Temperaturklassen neu berechnen, da diese NA Werte enthalten
```{r}
# Klasseneinteilung in kalt (Temp<10), normla (Temp10-20) und warm (Temp>20)
master_df$Temperaturklassen <- "kalt"
master_df$Temperaturklassen[master_df$Temperatur > 10 & master_df$Temperatur < 20] <- "normal"
master_df$Temperaturklassen[master_df$Temperatur > 20] <- "warm"

```


Na Value mit 0 ersetzten
```{r}
#master_df <- master_df %>%
#    mutate(KielerWoche = if_else(is.na(KielerWoche), 0, KielerWoche))

#NA Werte durch 0 ersetzen. 
master_df[is.na(master_df)] <- 0

#von Merle eingefügt, damit alle NAs mit 0 ersetzt werden  klappte gerstern aber nicht heute
#master_df[is.na(master_df)] <- 0 
```

Modellgleichung NEU
```{r}
mod1 <- lm(Gesamtumsatz ~ Ware_1, master_df)
mod2 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3, master_df)
mod3 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3 + Ware_4, master_df)
mod4 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3 + Ware_4 + Ware_5, master_df)
mod5 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6, master_df)
mod6 <- lm(Gesamtumsatz ~ Ware_1 + Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6 + as.factor(Temperaturklassen), master_df) #76% Testsieger
#mod6 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6 + Wochentag, master_df) #sehr kleine Verbesserung
mod7 <- lm(Gesamtumsatz ~ Ware_1 + Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6 + as.factor(Temperaturklassen), master_df) 
#mod7 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6 + Wochentag + Ferien, master_df) #Sehr kleine Verschlechterung

summary(mod1)

```

Modellgleichung 

```{r}
#sobald die Warengruppen mit einbezogen werden, steigt der adj r squared auf 99 oder 100
#
mod1 <- lm(Gesamtumsatz ~ Ware_2, master_df)
mod2 <- lm(Gesamtumsatz ~ Ware_1 + Ware_2 + Ware_3, master_df)
mod3 <- lm(Gesamtumsatz ~ Ware_2, master_df)
mod4 <- lm(Gesamtumsatz ~ Ware_1 + Ware_2 + Ware_3 + Ware_4 + Ware_5, master_df)
mod5 <- lm(Gesamtumsatz ~ Ware_1 + Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6, master_df)
mod6 <- lm(Gesamtumsatz ~ Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6 + as.factor(Temperaturklassen) + as.factor(vor_feiertag) + as.factor(Feiertag) + as.factor(Wochentag), master_df) 
#mod6 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6 + Wochentag, master_df) #sehr kleine Verbesserung
mod7 <- lm(Gesamtumsatz ~ Ware_2 + as.factor(Temperaturklassen) + as.factor(vor_feiertag) + as.factor(Feiertag) + as.factor(Wochentag), master_df)
#mod7 <- lm(Gesamtumsatz ~ Ware_1+ Ware_2 + Ware_3 + Ware_4 + Ware_5 + Ware_6 + Wochentag + Ferien, master_df) #Sehr kleine Verschlechterung

summary(mod7)

```


```{r}
# Preparation of Model Results

rbind(glance(mod1), glance(mod2), glance(mod3), glance(mod4), glance(mod5), glance(mod6), glance(mod7))
```


modelle plotten
```{r}
# with linear trend
p2 <- ggplot(master_df, aes(x=Gesamtumsatz, y=Datum)) +
  geom_point() +
  geom_smooth(method=lm , color="red", se=FALSE) +
  theme_ipsum()

#abline(mod7) #Add a regression line

#view(p2)



#plot(master_df$Gesamtumsatz, pch = 16, col = "blue")
#abline(mod6)
```
