==> .ABGABE ORDNER Steffen/Programmcode/7_Abbildungen.Rmd <==
---
title: "Abbildungen "
output: html_notebook
---


```{r}
#Bibliotheken
library(readr)
library(tidyverse)
library(lubridate)
library(ggplot2)
library(dplyr)

```





```{r}
# master df einlesen
master_df <- read.csv("master_df.csv")
#master_df[is.na(master_df)] <- 0 
master_df$Datum <- as.Date(master_df$Datum, format = "%Y-%m-%d")

#Variable mit dem ausgeschrieben Wochentag erzeugen (von Merle eingefügt)
#master_df$wochentag <- weekdays(master_df$Datum)
```


für die unterschiedlichen Monate ein Diagramme erstellen 


```{r}
#Einlesen der Daten

umsatz <- read_csv("umsatzdaten_gekuerzt.csv", 
    col_types = cols(Datum = col_datetime(format = " %Y-%m-%d   "))) #Eingeben des Datums

#Erstellung der Variable mit Jahr
umsatz$Monat <- month(umsatz$Datum, label = TRUE)




```

Tabelle und Abbildung  für den Umsatz der Warengruppen nach Monat
```{r}
#Tabelle für den Umsatz nach Warengruppen  und Monaten 
grafik_monat<- umsatz %>%
  group_by(Monat, Warengruppe) %>%
  summarise(
    n=n(), #ist praktisch, weil es die Anzahl z?hlt
    mean=mean(Umsatz),
    sd=sd(Umsatz),
    summe=sum(Umsatz)) %>%
  mutate( se=sd/sqrt(n))  %>%
  mutate( ic=se * qt((1-0.05)/2 + .5, n-1))


#Bild mit Koinfidenzintervall
# Confidence Interval
ggplot(grafik_monat) +
  geom_bar( aes(x=Monat, y=mean), stat="identity", fill="forestgreen", alpha=0.5) +
  geom_errorbar( aes(x=Monat, ymin=mean-ic, ymax=mean+ic), width=0.4, colour="orange", alpha=0.9, size=1.5) +
  xlab("Monat")+
  ylab("Durschnitt (€)")+
  facet_wrap(~Warengruppe)+
  ggtitle("Durchschnittlicher Umsatz im Monat über alle Warengruppen")


```


#Bild mit Koinfidenzintervallen vor und nach einem Feiertag erstellen
```{r}

#Datensatz auf die interessanten Variablen herunterbrechen 
Feiertage<-select(master_df, Datum, vor_feiertag, nach_feiertag, Feiertag)

#den Wert 1 durch das Statement ersetzten 
Feiertage<-Feiertage %>% 
  mutate(vor_feiertag = as.character(vor_feiertag)) %>% 
  mutate(vor_feiertag = replace(vor_feiertag, vor_feiertag == 1, 'vor_feiertag'))

Feiertage<-Feiertage %>% 
  mutate(nach_feiertag = as.character(nach_feiertag)) %>% 
  mutate(nach_feiertag = replace(nach_feiertag, nach_feiertag == 1, 'nach_feiertag'))


Feiertage<-Feiertage %>% 
  mutate(Feiertag = as.character(Feiertag)) %>% 
  mutate(Feiertag = replace(Feiertag, Feiertag == 1, 'Feiertag'))

#die Werte o durch NA ersetzten
Feiertage[Feiertage == 0] <- NA

#die Spalte in eine Reihe tun

Feiertage<-Feiertage %>% mutate(ftag = coalesce(vor_feiertag,nach_feiertag, Feiertag)) %>%
         select(Datum,ftag)


#beides nochmal als Datum definieren 
Feiertage$Datum <- as.Date(Feiertage$Datum, format = "%Y-%m-%d")
umsatz$Datum<- as.Date(umsatz$Datum, format = "%Y-%m-%d")

#einen left_join machen, um die beiden df`s zusammenzufügen 
Feiertage<-left_join(umsatz, Feiertage, by= "Datum")




#Tabelle für den Umsatz nach Warengruppen  und Feiertagen 
grafik_feiertag<- Feiertage %>%
  group_by(ftag, Warengruppe) %>%
  summarise(
    n=n(), #ist praktisch, weil es die Anzahl z?hlt
    mean=mean(Umsatz),
    sd=sd(Umsatz),
    summe=sum(Umsatz)) %>%
  mutate( se=sd/sqrt(n))  %>%
  mutate( ic=se * qt((1-0.05)/2 + .5, n-1))

#alle NAs rausschmeißen
grafik_feiertag<-na.omit(grafik_feiertag)

#Bild mit Koinfidenzintervall
# Confidence Interval
ggplot(grafik_feiertag) +
  geom_bar( aes(x=ftag, y=mean), stat="identity", fill="forestgreen", alpha=0.5) +
  geom_errorbar( aes(x=ftag, ymin=mean-ic, ymax=mean+ic), width=0.4, colour="orange", alpha=0.9, size=1.5) +
   ylab("Durchschnitt (€)")+
  xlab("Feiertage")+
  facet_wrap(~Warengruppe)+
  ggtitle("Umsätze vor, während und nach einem Feiertag")


```





#Bild mit Koinfidenzintervallen nach den Temperaturklassen erstellen

```{r}

#Datensatz auf die interessanten Variablen herunterbrechen 
temperatur<-select(master_df, Datum, Temperaturklassen )

#einen left_join machen, um die beiden df`s zusammenzufügen 
temperatur<-left_join(umsatz, temperatur, by= "Datum")

#Tabelle für den Umsatz nach Warengruppen  und Temperaturklassen


grafik_temperatur<- temperatur%>%
  group_by(Temperaturklassen, Warengruppe) %>%
  summarise(
    n=n(), #ist praktisch, weil es die Anzahl z?hlt
    mean=mean(Umsatz),
    sd=sd(Umsatz),
    summe=sum(Umsatz)) %>%
  mutate( se=sd/sqrt(n))  %>%
  mutate( ic=se * qt((1-0.05)/2 + .5, n-1))

#alle NAs rausschmeißen
grafik_temperatur<-na.omit(grafik_temperatur)


#Bild mit Koinfidenzintervall
# Confidence Interval
ggplot(grafik_temperatur) +
  geom_bar( aes(x=Temperaturklassen, y=mean), stat="identity", fill="forestgreen", alpha=0.5) +
  geom_errorbar( aes(x=Temperaturklassen, ymin=mean-ic, ymax=mean+ic), width=0.4, colour="orange", alpha=0.9, size=1.5) +
   ylab("Durchschnitt (€)")+
  xlab("Temperaturklasse")+
  facet_wrap(~Warengruppe)+
  ggtitle("Umsätze an Tagen mit kalter, normaler oder warmer Temperatur")
```
